backend/debug_schema.py

import os
import sys
from dotenv import load_dotenv
from supabase import create_client

load_dotenv()

supabase_url = os.getenv("SUPABASE_URL")
supabase_key = os.getenv("SUPABASE_KEY")

if not supabase_url or not supabase_key:
    print("Error: SUPABASE_URL or SUPABASE_KEY not found in environment.")
    sys.exit(1)

sb = create_client(supabase_url, supabase_key)

def check_structure():
    print("Checking matches table constraints...")
    
    # We can't easily query information_schema via PostgREST unless exposed.
    # But we can try to infer it by checking if we have a table that stores this info 
    # OR better: we can try to inspect how the 'matches' table behaves.
    
    # Since we can't run SQL, we will rely on the fact that the previous reproduction passed.
    # That implies the constraints *should* be fine for the test user.
    
    # However, let's try to verify if RLS is enabled.
    # There is no direct "is RLS enabled" check via standard client without SQL.
    
    print("Skipping direct schema query (restricted via API).")
    print("Generating SQL fix script instead...")

if __name__ == "__main__":
    check_structure()
