backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
slowapi==0.1.9

# Document Processing
python-docx==1.1.2
PyMuPDF==1.24.13

# NLP & AI
spacy==3.7.6
# Using standard sentence-transformers, rely on transitive deps
sentence-transformers==3.3.1

# Database
supabase==2.9.1

# Utilities
python-dotenv==1.0.1

# Development Tools
ruff==0.8.4
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
pre-commit==4.0.1
